+++ influxql/scanner.go
@@ -64,6 +64,15 @@ func (s *Scanner) Scan() (tok Token, pos Pos, lit string) {
 	case '*':
 		return MUL, pos, ""
 	case '/':
+		ch1, _ := s.r.read()
+		if ch1 == '*' {
+			if err := s.skipUntilEndComment(); err != nil {
+				return ILLEGAL, pos, ""
+			}
+			return COMMENT, pos, ""
+		} else {
+			s.r.unread()
+		}
 		return DIV, pos, ""
 	case '%':
 		return MOD, pos, ""
@@ -135,6 +144,36 @@ func (s *Scanner) scanWhitespace() (tok Token, pos Pos, lit string) {
 	}
 
 	return WS, pos, buf.String()
+}
+
+// skipUntilNewline skips characters until it reaches a newline.
+func (s *Scanner) skipUntilNewline() {
+	for {
+		if ch, _ := s.r.read(); ch == '\n' || ch == eof {
+			return
+		}
+	}
+}
+
+// skipUntilEndComment skips characters until it reaches a '*/' symbol.
+func (s *Scanner) skipUntilEndComment() error {
+	for {
+		if ch1, _ := s.r.read(); ch1 == '*' {
+			// We might be at the end.
+		star:
+			ch2, _ := s.r.read()
+			if ch2 == '/' {
+				return nil
+			} else if ch2 == '*' {
+				// We are back in the state machine since we see a star.
+				goto star
+			} else if ch2 == eof {
+				return io.EOF
+			}
+		} else if ch1 == eof {
+			return io.EOF
+		}
+	}
 }
 
 func (s *Scanner) scanIdent(lookup bool) (tok Token, pos Pos, lit string) {
@@ -230,6 +269,10 @@ func (s *Scanner) scanNumber() (tok Token, pos Pos, lit string) {
 		} else if ch == '+' {
 			return ADD, pos, ""
 		} else if ch == '-' {
+			if ch1 == '-' {
+				s.skipUntilNewline()
+				return COMMENT, pos, ""
+			}
 			return SUB, pos, ""
 		}
 	} else if ch == '.' {